requirements.txt
# Core Dependencies
sentence-transformers==2.2.2
qdrant-client==1.16.2
numpy==1.26.4
pandas==2.1.4

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# UI
streamlit==1.29.0
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
